extensionMonitor is a module now
