+++ src/background.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Background Page</title>
+    <script src="background.js" type="module"></script>
+  </head>
+  <body></body>
+</html>
