+++ src/manifest.json
@@ -15,7 +15,7 @@
   },
 
   "background": {
-    "scripts": ["background.js"]
+    "page": "background.html"
   },
 
   "browser_action": {
